requirements.txt
fastapi==0.104.1
python-dotenv==1.0.0
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.3
email-validator==2.1.0.post1

